Move `BrewingRecipe`
