+++ plugin/src/main/java/com/denizenscript/denizen/nms/interfaces/ItemHelper.java
@@ -5,12 +5,12 @@ import com.denizenscript.denizen.nms.util.jnbt.CompoundTag;
 import com.denizenscript.denizen.nms.util.jnbt.IntArrayTag;
 import com.denizenscript.denizen.nms.util.jnbt.Tag;
 import com.denizenscript.denizen.objects.ItemTag;
+import com.denizenscript.denizen.utilities.inventory.BrewingRecipe;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
 import org.bukkit.block.data.BlockData;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.RecipeChoice;
 import org.bukkit.inventory.ShapedRecipe;
 
 import java.util.List;
@@ -75,18 +75,6 @@ public abstract class ItemHelper {
     }
 
     public abstract boolean isValidMix(ItemStack input, ItemStack ingredient);
-
-    public static class BrewingRecipe {
-        public RecipeChoice input;
-        public RecipeChoice ingredient;
-        public ItemStack result;
-
-        public BrewingRecipe(RecipeChoice ingredient, RecipeChoice input, ItemStack result) {
-            this.ingredient = ingredient;
-            this.input = input;
-            this.result = result;
-        }
-    }
 
     public Map<NamespacedKey, BrewingRecipe> getCustomBrewingRecipes() {
         throw new UnsupportedOperationException();
